Remove connection factory - wip - remove Cluster's connection_factory
